Rename CXL Host entities
