+++ opencxl/bin/test_external_switch_cxl.py
@@ -1,9 +1,9 @@
 import asyncio
 from opencxl.cxl.environment import parse_cxl_environment
 from opencxl.apps.single_logical_device import SingleLogicalDevice
-from opencxl.apps.cxl_complex_host import (
-    CxlComplexHost,
-    CxlComplexHostConfig,
+from opencxl.apps.cxl_host import (
+    CxlHost,
+    CxlHostConfig,
     RootPortClientConfig,
     ROOT_PORT_SWITCH_TYPE,
     RootComplexMemoryControllerConfig,
@@ -36,7 +36,7 @@ class TestRunner:
     async def run_test(self):
         logger.info("Waiting for Apps to be ready")
         await self.wait_for_ready()
-        host = cast(CxlComplexHost, self._apps[0])
+        host = cast(CxlHost, self._apps[0])
         pci_bus_driver = PciBusDriver(host.get_root_complex())
         logger.info("Starting PCI bus driver init")
         await pci_bus_driver.init()
@@ -78,7 +78,7 @@ def main():
     # Add Host
     switch_host = "0.0.0.0"
     switch_port = 8000
-    host_config = CxlComplexHostConfig(
+    host_config = CxlHostConfig(
         host_name="CXLHost",
         root_bus=0,
         root_port_switch_type=ROOT_PORT_SWITCH_TYPE.PASS_THROUGH,
@@ -90,7 +90,7 @@ def main():
             memory_size=0x10000, memory_filename="memory_dram.bin"
         ),
     )
-    host = CxlComplexHost(host_config)
+    host = CxlHost(host_config)
     apps.append(host)
 
     # Add PCI devices
